coral_pitch_analysis/requirements.txt
streamlit>=1.28.0
elevenlabs>=0.2.0
python-dotenv>=1.0.0
requests>=2.31.0
pathlib2>=2.3.7
